WIP: Make the ColorMode of a RenderSettings editable in the UI.
TODO: fix the divergence of strategies for TrackSchemeStyle and
RenderSettings:
TrackSchemeStyles instance are unique and passed to the editor to be
modified. For RenderSettings, the editor works on a local copy and
copies the copy back to the list of 'true' RenderSettings.
The last one is harder to think about but probably better to avoid
dealing with listener deregistration etc...