Use the R-Tree in LineageTreeLayout.
FIXME Does not work for us:
1. A new tree instance is created each time we call add().
2. The edges are stored as if they were real objects. We
have to create a true R-Tree structure specialized for
Ref collections, as for the org.mastodon.kdtree.

However it is worth noting that no other part of the code is affected
by the change to a R-Tree, and that the R-Tree does not need to be
stored or accessed elsewhere. It might be handy to augment the lineage
class with methods that exploit it though for user interaction.