A Listenable graph wrapper that only exposes vertices marked as visible.
The visibility of vertices is controlled by a BooleanPropertyMap given at
construction.

However we return all edges. We do not care whether we return edges that
are connected non-visible vertices. In practice this will not impact
the application, lest a view chooses to query vertices through edges
instead of the opposite.